% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_curve.R
\name{gene_curve}
\alias{gene_curve}
\title{This function plot the gene trajectory from drosophila data at embryonic stage}
\usage{
gene_curve(input_ids, plot.only = TRUE)
}
\arguments{
\item{input_ids}{a vector contains either flybase ID, CG_ID or genesymbol; can be mixed}
}
\description{
This function plot the gene trajectory from drosophila data at embryonic stage
}
\examples{
\dontrun{
input_ids = c("CG11880", "tin", "CG6404","XXXXXXX","d4", "Traf6")
gene_curve(input_ids)
}
}

