% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_id.R
\name{matched_id}
\alias{matched_id}
\title{This function identity the flybase ID, CG_ID or genesymbol}
\usage{
matched_id(input_ids)
}
\arguments{
\item{input_ids}{a vector contains either flybase ID, CG_ID or genesymbol; can be mixed}
}
\description{
This function identity the flybase ID, CG_ID or genesymbol
}
\examples{
\dontrun{
input_ids = c("CG11880", "tin", "CG6404","XXXXXXX","d4", "Traf6")
matched_id(input_ids)
}
}

