% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeIDs.R
\name{mergeIDs}
\alias{mergeIDs}
\title{This function merge the flyID by either genesymbol, CG_ID or flybase ID}
\usage{
mergeIDs(input, merge_id_x = c("id1"), merge_id_y = c("CG_ID", "flybaseID",
  "genesymbol"))
}
\arguments{
\item{input}{data.frame to merge with}

\item{merge_id_x}{colname in input data frame to merge with}

\item{merge_id_y}{colname in the reference file, one of "CG_ID",  "flybaseID" , "genesymbol"}
}
\description{
This function merge the flyID by either genesymbol, CG_ID or flybase ID
}
\examples{
\dontrun{
input = structure(list(id1 = c("FBgn0011656", "FBgn0011656", "FBgn0011656", "FBgn0011656", "FBgn0011656", "FBgn0011656"), id2 = c("FBgn0004009", "FBgn0004110", "FBgn0010246", "FBgn0039039", "FBgn0086906", "FBgn0264491")), .Names = c("id1", "id2"), row.names = c(NA, -6L), class = "data.frame")
input\%>\%mergeIDs(input=.,merge_id_x="id1",merge_id_y="flybaseID")\%>\%mergeIDs(input=.,merge_id_x="id2",merge_id_y="flybaseID")\%>\%arrange(id2,type)\%>\%filter(type!="rna")
}
}

