% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regfun_slos.R
\name{regfun_slos}
\alias{regfun_slos}
\title{This function compute the regulation functions via slos penalty}
\usage{
regfun_slos(xfdlist, yfd, Lcholy, time_obs = seq(0, 23, len = 24), yname,
  xnames, lambda = 0.001, gamma = 0.001, maxiteration = 500,
  lambdaI = 1e+05, d = 5, K = 5, maxabs = 1e-06, verbose = TRUE,
  type = "derivatives")
}
\arguments{
\item{xfdlist}{a list with each element being a fd object, for all the predictors}

\item{yfd}{a fd object for the response}

\item{time_obs}{evaluated time points}

\item{yname}{reponse name}

\item{xnames}{matches the xfdlist}

\item{lambda}{the fscad penalty parameter}

\item{gamma}{the smoothing parameter}

\item{d}{the order}

\item{K}{the number of inside knots}
}
\description{
This function compute the regulation functions via slos penalty
}
\examples{
\dontrun{
yfd = fdlist[[14]]
xfdlist  = fdlist
time_obs = (eval_times = timepoints$realtime)
xnames = names(xfdlist)
res = regfun_slos(xfdlist,yfd,time_obs,yname=xnames[14],xnames)
}
}

