% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_each.R
\name{search_each}
\alias{search_each}
\title{This function identity search the interactions in the flybase data}
\usage{
search_each(s, input_interactions = input_int)
}
\arguments{
\item{s}{interaction type one of "ppi","rna","tf","generic"}

\item{input_interactions}{data.frame with two cols}
}
\description{
This function identity search the interactions in the flybase data
}

