% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_interactions.R
\name{search_interaction}
\alias{search_interaction}
\title{This function identity search the interactions in the flybase data}
\usage{
search_interaction(input_interactions, type = c("all", "ppi", "rna", "tf",
  "generic"))
}
\arguments{
\item{input_interactions}{data.frame with two cols}

\item{type}{interaction type one of 'all',"ppi","rna","tf","generic"}
}
\description{
This function identity search the interactions in the flybase data
}
\examples{
\dontrun{
input_interactions = data.frame(id1 = c('FBgn0011656','FBgn0003149',"7872"),id2 = c('FBgn0003149','FBgn0003900',"9232"))
output = search_interaction(input_interactions=pairs,type="all")
output\%>\%do(data.frame(type=.$type,.$outdata\%>\%select(id1,id2)))\%>\%group_by(type)\%>\%summarise(n=n())
output\%>\%do(data.frame(type=.$type,.$outdata))\%>\%select(id1,id2,PMID_URL)\%>\%tail
}
}

